bookCollapseSection to false
Nested lists are badly recognized by screen reader (The list of topics on the left side (read, organize …) may be organized as a nested list. Currently, the screen reader reads “right pointing small triangle”   after each menu item and it does not indicate if the item is expandable or not. )
A better solution could be found to insure nested lists are nested lists.